fox/collectorfox/requirements.txt
Flask==3.0.3
lance==1.2.1
lancedb==0.14.0
pyarrow==17.0.0
Requests==2.32.3
tensorflow==2.17.0
tensorflow_datasets==4.9.6
fastembed==0.4.1
open_clip_torch
pillow